gui/requirements.txt
dearpygui==2.0.0
numpy==2.3.1
pandas==2.3.0
pyserial==3.5
PyYAML==6.0.2
PyYAML==6.0.2
